+++ typescript/src/lib/contracts/wallet-registry.ts
@@ -1,10 +1,7 @@
 import { Hex } from "../utils"
-import {
-  DkgResultApprovedEvent,
-  DkgResultChallengedEvent,
-  DkgResultSubmittedEvent,
-} from "../../wallet"
-import { GetEvents } from "./chain-event"
+import { Event, GetEvents } from "./chain-event"
+import { BigNumber } from "ethers"
+import { Identifier } from "./chain-identifier"
 
 /**
  * Interface for communication with the WalletRegistry on-chain contract.
@@ -34,4 +31,99 @@ export interface WalletRegistry {
    * @see GetEventsFunction
    */
   getDkgResultChallengedEvents: GetEvents.Function<DkgResultChallengedEvent>
+}
+
+/**
+ * Represents an event emitted when a DKG result is submitted to the on-chain
+ * wallet registry.
+ */
+export type DkgResultSubmittedEvent = {
+  /**
+   * 32-byte hash of the submitted DKG result.
+   */
+  resultHash: Hex
+  /**
+   * 32-byte seed of the current DKG execution.
+   */
+  seed: Hex
+  /**
+   * DKG result object.
+   */
+  result: DkgResult
+} & Event
+
+/**
+ * Represents an event emitted when a DKG result is approved on the on-chain
+ * wallet registry.
+ */
+export type DkgResultApprovedEvent = {
+  /**
+   * 32-byte hash of the submitted DKG result.
+   */
+  resultHash: Hex
+  /**
+   * Approver's chain identifier.
+   */
+  approver: Identifier
+} & Event
+
+/**
+ * Represents an event emitted when a DKG result is challenged on the on-chain
+ * wallet registry.
+ */
+export type DkgResultChallengedEvent = {
+  /**
+   * 32-byte hash of the submitted DKG result.
+   */
+  resultHash: Hex
+  /**
+   * Challenger's chain identifier.
+   */
+  challenger: Identifier
+  /**
+   * Reason of the challenge.
+   */
+  reason: string
+} & Event
+
+/**
+ * Represents a DKG on-chain result.
+ */
+type DkgResult = {
+  /**
+   * Claimed submitter candidate group member index. Is in range [1, groupSize].
+   */
+  submitterMemberIndex: BigNumber
+  /**
+   * Generated group public key.
+   */
+  groupPubKey: Hex
+  /**
+   * Array of misbehaved members indices (disqualified or inactive). Indices
+   * are in range [1, groupSize], unique, and sorted in ascending order.
+   */
+  misbehavedMembersIndices: number[]
+  /**
+   * Concatenation of signatures from members supporting the result.
+   * The message to be signed by each member is keccak256 hash of the
+   * calculated group public key, misbehaved members indices and DKG
+   * start block. The calculated hash is also prefixed with
+   * `\x19Ethereum signed message:\n` before signing.
+   */
+  signatures: Hex
+  /**
+   * Indices of members corresponding to each signature. Indices are
+   * in range [1, groupSize], unique, and sorted in ascending order.
+   */
+  signingMembersIndices: BigNumber[]
+  /**
+   * Identifiers of candidate group members as outputted by the group
+   * selection protocol.
+   */
+  members: number[]
+  /**
+   * Keccak256 hash of group members identifiers that actively took part
+   * in DKG (excluding IA/DQ members).
+   */
+  membersHash: Hex
 }
